move getVariant call to background thread

+++ experiments/experiments-api/src/main/java/com/duckduckgo/experiments/api/VariantConfig.kt
@@ -18,7 +18,7 @@ package com.duckduckgo.experiments.api
 
 data class VariantConfig(
     val variantKey: String,
-    val weight: Float? = 0.0f,
+    val weight: Double? = 0.0,
     val filters: VariantFilters? = null,
 )
